Update to slang-v2.3.3
This version fixes an issue with some console applications (like
Midnight Commander) that have issue displaying colors and handle key
codes for the latest version of the xterm-256color terminfo.

This disables the patches from the el7 specfile. The perm patch is
already included in v2.3.3 and the aarch64 support patch is not needed
for XCP-ng build.

Signed-off-by: Thierry Escande <thierry.escande@vates.tech>